feat: 调整Watchdog.adapters数据结构，简化层次方便遍历

+++ watchdog/handler/file.go
@@ -11,7 +11,7 @@ type FileAdapter struct {
 	Name     string
 	Config   *FileAdapterCfg
 	logger   log.Logger
-	Priority int
+	Priority uint8
 }
 
 type FileAdapterCfg struct {
